+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Ongoing development.
 
 * generalize `decomp_nil_eq_elt` and `decomp_unit_eq_elt` to `decomp_nil_eq` and `decomp_unit_eq`
 * rename `list_to_Forall` into `list_to_ForallT`, `Forall_to_list` into `ForallT_to_list`, etc.
+* correct bad namings in `shuffle.v` and `shuffleT.v`
+  * rename `sublist_*` into `shuffle[T]_*`
+  * exchange names of `Permutation[T]_shuffle[T]` and `shuffle[T]_Permutation[T]`
 * weaken uselessly strong hypothesis in `finite_strictorder_max`
 
 ## Version 2.1.0